resume finished downloads (fix #12109)
