+++ main/src/cgeo/geocaching/utils/DebugUtils.java
@@ -8,9 +8,6 @@ import cgeo.geocaching.ui.TextParam;
 import cgeo.geocaching.ui.dialog.SimpleDialog;
 
 import android.app.Activity;
-import android.app.DownloadManager;
-import android.content.Context;
-import android.database.Cursor;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Handler;
@@ -84,86 +81,6 @@ public class DebugUtils {
                     (dialog, which) -> createLogcatHelper(activity, true, false, null)
                 );
         }
-    }
-
-    public static void dumpDownloadmanagerInfos(@NonNull final Activity activity) {
-        final DownloadManager downloadManager = (DownloadManager) activity.getSystemService(Context.DOWNLOAD_SERVICE);
-        final DownloadManager.Query query = new DownloadManager.Query();
-        final StringBuilder sb = new StringBuilder();
-        try (Cursor c = downloadManager.query(query)) {
-            final int columnStatus = c.getColumnIndex(DownloadManager.COLUMN_STATUS);
-            final int columnReason = c.getColumnIndex(DownloadManager.COLUMN_REASON);
-            final int[] columns = {
-                c.getColumnIndex(DownloadManager.COLUMN_ID),
-                c.getColumnIndex(DownloadManager.COLUMN_TITLE),
-                columnStatus,
-                columnReason,
-                c.getColumnIndex(DownloadManager.COLUMN_BYTES_DOWNLOADED_SO_FAR),
-                c.getColumnIndex(DownloadManager.COLUMN_TOTAL_SIZE_BYTES),
-                c.getColumnIndex(DownloadManager.COLUMN_LAST_MODIFIED_TIMESTAMP),
-                c.getColumnIndex(DownloadManager.COLUMN_MEDIA_TYPE),
-                c.getColumnIndex(DownloadManager.COLUMN_URI),
-                c.getColumnIndex(DownloadManager.COLUMN_MEDIAPROVIDER_URI),
-                c.getColumnIndex(DownloadManager.COLUMN_LOCAL_URI)
-            };
-            while (c.moveToNext()) {
-                for (int column : columns) {
-                    sb.append("- ").append(c.getColumnName(column)).append(" = ");
-                    if (column == columnStatus) {
-                        sb.append(c.getString(column));
-                        final int status = c.getInt(column);
-                        if (status == DownloadManager.STATUS_FAILED) {
-                            sb.append(" - download has failed (and will not be retried)");
-                        } else if (status == DownloadManager.STATUS_PAUSED) {
-                            sb.append(" - download is waiting to retry or resume");
-                        } else if (status == DownloadManager.STATUS_PENDING) {
-                            sb.append(" - download is waiting to start");
-                        } else if (status == DownloadManager.STATUS_RUNNING) {
-                            sb.append(" - download is currently running");
-                        } else if (status == DownloadManager.STATUS_SUCCESSFUL) {
-                            sb.append(" - download has successfully completed");
-                        }
-                    } else if (column == columnReason) {
-                        final int reason = c.getInt(column);
-                        sb.append(reason);
-                        if (reason == DownloadManager.PAUSED_QUEUED_FOR_WIFI) {
-                            sb.append(" - paused: download exceeds a size limit for downloads over mobile network / waiting for Wifi");
-                        } else if (reason == DownloadManager.PAUSED_UNKNOWN) {
-                            sb.append(" - paused: for some other reason");
-                        } else if (reason == DownloadManager.PAUSED_WAITING_FOR_NETWORK) {
-                            sb.append(" - paused: waiting for network connectivity to proceed");
-                        } else if (reason == DownloadManager.PAUSED_WAITING_TO_RETRY) {
-                            sb.append(" - paused: some network error occurred and the download manager is waiting before retrying the request");
-                        } else if (reason == DownloadManager.ERROR_CANNOT_RESUME) {
-                            sb.append(" - error: some possibly transient error occurred but we can't resume the download");
-                        } else if (reason == DownloadManager.ERROR_DEVICE_NOT_FOUND) {
-                            sb.append(" - error: no external storage device was found. SD card mounted?");
-                        } else if (reason == DownloadManager.ERROR_FILE_ALREADY_EXISTS) {
-                            sb.append(" - error: requested destination file already exists, will not be overwritten");
-                        } else if (reason == DownloadManager.ERROR_FILE_ERROR) {
-                            sb.append(" - error: unknown storage issue");
-                        } else if (reason == DownloadManager.ERROR_HTTP_DATA_ERROR) {
-                            sb.append(" - error: HTTP processing error at data level");
-                        } else if (reason == DownloadManager.ERROR_INSUFFICIENT_SPACE) {
-                            sb.append(" - error: insufficient storage space");
-                        } else if (reason == DownloadManager.ERROR_TOO_MANY_REDIRECTS) {
-                            sb.append(" - error: too many redirects");
-                        } else if (reason == DownloadManager.ERROR_UNHANDLED_HTTP_CODE) {
-                            sb.append(" - error: unhandled HTTP cod");
-                        } else if (reason == DownloadManager.ERROR_UNKNOWN) {
-                            sb.append(" - error: unknown error");
-                        }
-
-                    } else {
-                        sb.append(c.getString(column));
-                    }
-                    sb.append("\n");
-                }
-                sb.append("\n---\n\n");
-            }
-        }
-
-        SimpleDialog.of(activity).setTitle(R.string.debug_current_downloads).setMessage(TextParam.text(sb.toString()).setMarkdown(true)).show();
     }
